requirements.txt
aiohttp==3.7.4.post0
arrow==1.2.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.3.0
canvasapi==3.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
colorama==0.4.6
cryptography==36.0.1
decorator==4.4.2
discord.py==1.7.3
distlib==0.3.4
filelock==3.4.2
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.0
google-cloud-storage==2.7.0
google-cloud-texttospeech==2.14.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==1.1.2
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.13.0
httplib2==0.21.0
idna==3.3
imageio==2.26.1
imageio-ffmpeg==0.4.8
instagrapi==1.16.30
moviepy==1.0.3
multidict==6.0.2
mutagen==1.46.0
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.1.0
pandas==1.5.3
Pillow==9.3.0
platformdirs==2.4.1
playwright==1.18.1
proglog==0.1.10
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.10.2
pyee==9.0.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.1.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
stripe==5.3.0
TikTokApi==4.1.0
tqdm==4.65.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.4.0
uptime==3.0.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
websockets==10.1
wsproto==1.0.0
yarl==1.7.2
